[DON'T MERGE] Extract Footer from main application controller
- add `footer.fxml`
- add `FooterController` and move code related to the footer area with
social media buttons
- share UI code via `FXUtils`

Part of #462
