+++ src/main/java/org/terasology/launcher/gui/javafx/FXUtils.java
@@ -1,0 +1,24 @@
+package org.terasology.launcher.gui.javafx;
+
+import javafx.animation.ScaleTransition;
+import javafx.scene.Node;
+import javafx.util.Duration;
+
+public class FXUtils {
+    /**
+     * Creates a {@link javafx.animation.ScaleTransition} with the given factor for the specified node element.
+     *
+     * @param factor the scaling factor
+     * @param node   the target node
+     * @return a transition object
+     */
+    public static ScaleTransition createScaleTransition(final double factor, final Node node) {
+        final ScaleTransition scaleTransition = new ScaleTransition(Duration.millis(200), node);
+        scaleTransition.setFromX(node.getScaleX());
+        scaleTransition.setFromY(node.getScaleY());
+        scaleTransition.setToX(factor);
+        scaleTransition.setToY(factor);
+        return scaleTransition;
+    }
+
+}
