+++ src/main/resources/org/terasology/launcher/views/application.fxml
@@ -21,7 +21,6 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
 <?import javafx.scene.control.ProgressBar?>
 <?import javafx.scene.control.Tab?>
 <?import javafx.scene.control.TabPane?>
@@ -34,7 +33,6 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
 <?import javafx.scene.web.WebView?>
 <?import java.lang.String?>
 <?import java.net.URL?>
@@ -56,89 +54,7 @@
         </ImageView>
         <BorderPane prefHeight="529.0" prefWidth="756.0">
           <bottom>
-            <StackPane maxHeight="-Infinity" prefHeight="64.0" prefWidth="200.0">
-              <children>
-                <HBox fx:id="bottomPanel" alignment="CENTER" maxHeight="-Infinity" prefHeight="64.0" prefWidth="979.0" spacing="8.0" style="-fx-background-color:  rgba(10,10,10,0.7);&#10;-fx-background-radius: 10 10 0 0;" StackPane.alignment="BOTTOM_CENTER">
-                  <children>
-                    <Button id="" fx:id="facebookButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openFacebook" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/facebook.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                    <Button id="" fx:id="githubButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openGithub" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/github.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                    <Button id="" fx:id="discordButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openDiscord" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/discord.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                    <Button id="" fx:id="redditButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openReddit" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/reddit.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                    <Button id="" fx:id="twitterButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openTwitter" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/twitter.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                    <Button id="" fx:id="youtubeButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openYoutube" onMouseEntered="#handleSocialButtonMouseEntered" onMouseExited="#handleSocialButtonMouseExited" onMousePressed="#handleSocialButtonMousePressed" onMouseReleased="#handleSocialButtonMouseReleased" styleClass="socialButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/youtube.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                    </Button>
-                  </children>
-                </HBox>
-                <HBox alignment="CENTER_LEFT" maxHeight="-Infinity" maxWidth="-Infinity" prefHeight="64.0" prefWidth="64.0" spacing="8.0" StackPane.alignment="BOTTOM_LEFT">
-                  <HBox.margin>
-                      <Insets left="8.0" />
-                  </HBox.margin>
-                  <children>
-                    <Button id="" fx:id="warningButton" contentDisplay="GRAPHIC_ONLY" graphicTextGap="0.0" mnemonicParsing="false" onAction="#openLogs" styleClass="warningButton">
-                      <graphic>
-                        <ImageView fitHeight="32.0" fitWidth="32.0" pickOnBounds="true" preserveRatio="true">
-                          <image>
-                            <Image url="@../images/warn.png" preserveRatio="false" smooth="false" />
-                          </image>
-                        </ImageView>
-                      </graphic>
-                  </Button>
-                  </children>
-                </HBox>
-                <Label fx:id="versionInfo" text="2.0.0-rc1" textFill="#dfdfdf" StackPane.alignment="BOTTOM_RIGHT">
-                  <font>
-                    <Font name="System Bold" size="12.0" />
-                  </font>
-                </Label>
-              </children>
-            </StackPane>
+            <fx:include fx:id="footer" source="footer.fxml"/>
           </bottom>
           <center>
             <AnchorPane prefHeight="200.0" prefWidth="200.0">
